feat(1): multifile environments - copy and download
